requirements.txt
disnake
asyncio
pytz
python-dotenv
numdeclination